refactor(money): represent currency as Decimal, not float
Eliminates IEEE-754 rounding error in monetary arithmetic — most importantly
the month-by-month debt snowball/avalanche simulation, which accumulated error
over up to 600 iterations.

- money.py: new helpers (to_decimal, optional_decimal, cents, ZERO/CENT).
- db.py: register a sqlite3 adapter so Decimal binds to REAL columns. Storage
  stays REAL (no schema migration); cent-quantized values round-trip exactly,
  and existing databases load unchanged.
- models: all currency/rate/share fields are now Decimal.
- repositories: coerce money to Decimal at the read boundary (to_decimal /
  optional_decimal); writes bind Decimal directly via the adapter.
- snowball.py: simulate entirely in Decimal; quantize to cents at the output
  boundary. extra/lump-sum inputs are coerced, so float callers still work.
- budget.py: monthly figures computed with exact Decimal division from integer
  (paychecks, months) ratios instead of pre-rounded float factors.
- dialogs/views: read model Decimals into spinboxes via float(), build models
  with cents()/to_decimal(), and coerce the fetched (float) stock price to
  Decimal before arithmetic. Decimal formats with the existing f-strings.

Verified: full Decimal unit tests, an offscreen GUI smoke test driving the
snowball/salary/stocks views, and a read-only load of the live finance.db all
pass; nix build succeeds.

+++ src/financeguru/models/debt.py
@@ -1,11 +1,12 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
 class Debt:
     id: int
     name: str
-    balance: float
-    interest_rate: float   # APR as a percentage, e.g. 18.5
-    minimum_payment: float
+    balance: Decimal
+    interest_rate: Decimal   # APR as a percentage, e.g. 18.5
+    minimum_payment: Decimal
     notes: str | None
