+++ src/financeguru/models/stock.py
@@ -1,12 +1,13 @@
 from dataclasses import dataclass
+from decimal import Decimal
 from typing import Optional
 
 
 @dataclass
 class Stock:
     ticker: str
-    shares: float
-    purchase_price: float
+    shares: Decimal
+    purchase_price: Decimal
     purchase_date: str
     notes: Optional[str] = None
     id: Optional[int] = None
