+++ src/financeguru/models/stock_tip.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -6,11 +7,11 @@ class StockTip:
     id: int
     ticker: str
     action: str
-    target_price: float | None
+    target_price: Decimal | None
     confidence: int
     notes: str | None
     added_date: str
     analyst_action: str | None = None
-    analyst_target: float | None = None
+    analyst_target: Decimal | None = None
     analyst_count: int | None = None
     analyst_updated: str | None = None
